pyproject.toml
[tool.poetry]
name = "smart-talk"
version = "0.1.0"
description = ""
authors = ["al-ov73 <klever1987@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.1"
jinja2 = "^3.1.4"
fastapi = "^0.115.4"
aiohttp = "^3.10.10"
gunicorn = "^23.0.0"
uvicorn = "^0.32.0"
python-multipart = "^0.0.17"
gigachat = "^0.1.35"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
